fix: consolidate rito type related traits
