+++ crates/ltk_ritobin/src/typecheck.rs
@@ -1,4 +1,1 @@
-mod name_ext;
-pub use name_ext::*;
-
 pub mod visitor;
